See description
I added the arrayLists to represent thee panel buttons. I added a start button.  I had to add some code to make the color change possible for the buttons. I have started to add action listeners to the panel buttons so that they can change the color of their neighbors.